Extract PictureThumbnails concern
Holds everything you need to make your model being able to
display image thumbnails and make use of the image cropper.

Your model needs to have `crop_from` and `crop_size` attributes
as well as a `Alchemy::Picture` association named `picture` and
a `settings` Hash.
