+++ spec/rails_helper.rb
@@ -21,6 +21,7 @@ require "alchemy/test_support"
 require "alchemy/test_support/config_stubbing"
 require "alchemy/test_support/essence_shared_examples"
 require "alchemy/test_support/having_crop_action_examples"
+require "alchemy/test_support/having_picture_thumbnails_examples"
 require "alchemy/test_support/integration_helpers"
 require "alchemy/test_support/shared_contexts"
 require "alchemy/test_support/shared_uploader_examples"
